Reord watchlist cmds into new package
